requirements.txt
# requirements.txt

# The main library that orchestrates browser interaction + LLM instructions
browser-use==0.1.29

# Used by browser-use under the hood for actual browser automation
playwright>=1.49.0

# Loads environment variables (like OPENAI_API_KEY) from a .env file
python-dotenv>=1.0.0

# Only needed if you plan on creating GIFs from screenshots
pillow>=11.1.0
